Set unused ReadHandles in TrackParticleCnvAlg to null in MuonStandalone configuration
TrackParticleCnvAlg is used here to convert MuonSpectrometer Trk::Track objects to xAOD::TrackParticle objects.
However, it can also be used to convert old-style Rec::TrackParticles to xAOD::TrackParticles.
To avoid creating unnecessary data dependencies, the handles for the latter, unused, process are now explicitly set to null in the configuration.
This addresses ATLASRECTS-4312.


Former-commit-id: 5790c06064787983776082b8ffc011ea7da68748