Replaced direct storegate access with data handles.
Also turned algorithms into a reentrant algorithm and removed the
properties to enable the Reta and and hadronic energy check. The latter
two optional cluster selection strategies are disabled by setting the
names of the corresponding tools to an empty string rather than
additonal boolean properties.

This addresses part of ATLASRECTS-4307.


Former-commit-id: 964a6fdfde7f0022d33fdb3b2b1ef712ad3dac41