Simplification of FCS_StepInfoSDs
After discussion on ATLASSIM-3538 it was decided to switch to using
a common simplified version of the `update_map` method in the
`FCS_StepInfoSD` base-class rather than re-implementing it in each
of the derived classes. The cleanest way to do this was to make
the time and radius cuts for merging arguments of the method.
The output should not be altered by these changes.


Former-commit-id: 443845635dfce5b45fe147f582f239e4f8b2ea02