AthenaKernel: Return range from CondCont by pointer.
The find operations in CondCont can optionally return the range
of the found element.  This was being copied into the ReadCondHandle.
This turned out to be a noticable overhead, especially since in most
cases, this information is not actually used.

Changed so that we instead store a pointer to the range in the
ReadCondHandle.


Former-commit-id: 390548c931e09fd51dfdaa567bf7aae49cd04c9e