iPatInterfaces: const fixes
DataVector has had a long-standing bug in which it was possible to get
a non-const pointer from a const_iterator.  These packages were relying
on this bug in order to compile.  Fix this in order to be able to finally
fix the underlying DataVector bug.

As part of this, we need to change IPointManager::spacePoints() and
ITrackFinder::trackSearch() so that the space point list is a
vector<const Trk::SpacePoint*>, rather than vector<Trk::SpacePoint*>.

There is still one place in iPatRecAlgs that abuses this bug in order to modify
objects already recorded in SG.  This will be addressed in a subsequent MR.



Former-commit-id: 61b2e8ddf863e4f7706632ff9a7fb4682bcc76bc