Only set data handle keys for existing data that is needed.
The data handle keys for the old AOD track particles are only set in
case they are to be converted to xAOD track particles and the track keys
are only set if xAOD track particles are to be created from tracks.
Moreover truth keys are only set for MC.

This solves unmet data dependencies discussed in ATLASRECTS-4312.


Former-commit-id: 83807b017b607af5a9d8bafdd93ba45268eac921