Remove some NSW data dependencies from MuonPRD_MultiTruthMaker and MuonTruthDecorationAlg
Prior to the advent of multithreading the code would simply check for the presence of NSW data collections and skip them if they were not present.
With multithreading, this approach is of course no longer viable.
MuonPRD_MultiTruthMaker now has a flag to control the use of NSW collections (and also CSC collections).
For the MuonTruthDecorationAlg, the PRD truth map and SDO collection names are each controlled by a single property, so I just changed the defaults to exclude the NSW.
Ideally there would be a consistent method for doing this across multiple truth algs but that's a secondary priority.
Also, I noticed that for some reason the name of the muon collection in the MuonTruthAssociationAlg was given as "muons" rather than "Muons", this has been fixed.
This will resolve the problem reported in ATLASRECTS-4323.


Former-commit-id: c3a97ff8816dfaf7b2762e7340c2c1c4f7bae8eb