Stop scheduling muon AOD to xAOD conversion algorithms
Due to a bug in the configuration algorithms for converting old-style AOD trackParticles, muons, and segments were being scheduled in a number of reconstruction steps.
This introduced unnecessary data dependencies and was responsible for the bug reported in ATLASRECTS-4298.
This problem is resolved by removing the scheduling of these algorithms from the main muon configuration entirely, as they are completely unnecessary in almost all cases.
If desired, the file MuonAODto_xAODCnv_jobOptions.py can be added to the configuration by hand.


Former-commit-id: de0dbe3f9586eb39e7fb80d2ab648328a188a4be