Use new method to retrieve the public track particle creator tool.
In the process of solving unmet data dependencies a method was
introduced to get a track particle creator tool rather than getting a
pre-configured tool from the global namespace. To manipulate the
configuration of the creator algorithm it is now accessed through the
top sequemnce rather than using a now non-existing variable from the
global namespace.


Former-commit-id: b147c7b50a88941bbc9e570e7c612e50110eb740