ALFA_BeamTransport: const workaround.
ALFA_BeamTransport is abusing a DataVector bug to alter constant objects
in StoreGate.

A long-standing bug in DataVector has meant that it was possible to get
a non-const pointer back from a DataVector::const_iterator.

ALFA_BeamTransport retrieves a MCEventCollection from SG and then
modifies it in TransportSelectedParticle.

This of course won't work with AthenaMT.

In preparation for trying to finally remove this bug from DataVector,
this has been adjusted so that the const_cast is explicit.


Former-commit-id: 533ebd11ab9293ea1b1f0f8c35d3d83a0cdbd767