ForwardTransportFast: const workaround.
ForwardTransportFast is abusing a DataVector bug to alter constant objects
in StoreGate.

A long-standing bug in DataVector has meant that it was possible to get
a non-const pointer back from a DataVector::const_iterator.

ForwardTransportFast retrieves a MCEventCollection from SG and then
modifies it in execute().

This of course won't work with AthenaMT.

In preparation for trying to finally remove this bug from DataVector,
this has been adjusted so that the const_cast is explicit.


Former-commit-id: 23bad0c4a611f700b8559bb5c4a9bc9c23ab6054