ForwardTransport: const workaround.
ForwardTransport is abusing a DataVector bug to alter constant objects
in StoreGate.

A long-standing bug in DataVector has meant that it was possible to get
a non-const pointer back from a DataVector::const_iterator.

ForwardTransport retrieves a GenEvent from SG (by way of
ForwardTransportSvc::getEvent()) and then modifies it in
ForwardTransportSvc::addMcVertex().

This of course won't work with AthenaMT.

In preparation for trying to finally remove this bug from DataVector,
this has been adjusted so that the const_cast is explicit.


Former-commit-id: 605ef5aebb2bf8811d06869460b7c88573c356d7