HepMCAnalysis_i: Const workarounds
A long-standing bug in DataVector has meant that it was possible to get
a non-const pointer back from a DataVector::const_iterator.
HepMCAnalysis is abusing this to get a non-const pointer to a HepMC::GenEvent
object in StoreGate.

With this non-const pointer, HepMCAnalysis calls MeVToGeV, changing all the numbers
in the object to non-standard units.  This needs to be changed.  Besides just
the constness issue, this will cause bugs if combined with anything else
that accesses GenEvent.  And all numbers in ATLAS software should use ATLAS
unit conventions.  If the units must be changed, you should make a copy of the
GenEvent object and change that.

But a non-const pointer is also required to pass to the Process() and ClearEvent()
methods of baseAnalysis.  However, from a spot check, it doesn't appear that these
actually do change the GenEvent.  If so, then these interfaces should really
be changed to take a const pointer.

For now, make this implicit conversion explicit with a const_cast.
This will be required in order to finally fix the underlying DataVector bug.




Former-commit-id: 66e186f82d6002e7bd4048c14292f23776d71558