TrigFTKSim: Const workaround.
A long-standing bug in DataVector has meant that it was possible to get
a non-const pointer back from a DataVector::const_iterator.

FTKMergerAlgo is abusing this in order to get non-const pointers to
FTKAthTrack objects retrieved from SG.  It then can modify these objects
(for example, by calling setHWRejected() and setHWTrackID() in removeTrack()).

This of course won't work for AthenaMT.

For the time being, the implicit cast has been replaced with an explicit
const_cast, in order to allow finally fixing the underlying DataVector bug.


Former-commit-id: f780e54fa5e5508de0613a3cc2e79822432e6ee7