Muon EF AthenaMT job option update (ATR-17553)
Updated the options so that if doEFSA is true, things will run without crashing.
At the moment we can't run both L2SA and EFSA at the same time (I think because we end up trying to decode the data twice... This part of the job options is still a bit of a mess, but my attempts to clean it up and keep both L2 and EF running were unsuccessful), so will leave doEFSA=False as default.


Former-commit-id: 1e832eabf2a3e83a2728118d023bd5de074bebcf