APR: retry PFC lookup to check if filename given is not an LFN
if Athena tries to open a file using a name that is registered in the
PFC as a LFN, but pretending it was an PFN, the PFC will find the entry
but not the file tech.  In this case, mark the name as LFN and retry the
lookup.
This makes it easy to use LFNs as input in Athena.


Former-commit-id: 25203dd29cde847be73ec035b0ce4be20291a042