iPatRecAlgs: Const workaround.
A long-standing bug in DataVector has meant that it was possible to get
a non-const pointer back from a DataVector::const_iterator.

iPatTrackTruthAssociator is abusing this in order to get non-const pointers
to Track* objects from StoreGate.  It then modifies these objects
by calling truthAssociation().
This won't work with AthenaMT.

For the time being, the implicit cast has been replaced with an explicit
const_cast, in order to allow finally fixing the underlying DataVector bug.


Former-commit-id: 840935cfadcbb40d329ea69747eb013f1ded273f