Fixing an ordering issue in particle jet definitions
As (I think?) described in ATLASG-1375, the current version of
CopyTruthJetParticles has an order dependence when the DR version of
photon dressing is used: the prompt leptons must appear before the FSR
photons in the truth record.  This commit changes the structure
somewhat, so that in a first loop through the truth record all prompt
leptons are identified, and in a second pass the FSR photons are
checked.  Should solve the issue, at the price of a small increase to
CPU.


Former-commit-id: 65b396ac0ff1d25d89499cb7a939131e84120ad6