JiveXML: Fix gcc8 warnings.
The function xdr_void in xdr.h is defined inconsistently.
An xdr procedure should take at least one argument; however, xdr_void does
not actually use its argument, so it is declared with none.
This causes gcc8 to give a warning when xdr_void gets cast to xdrproc_t
in clnt_call.

Presumably, this will get fixed upstream once gcc8 is released, but we can expect
it to be quite a while before we're building against that.  For now, explicitly
suppress these warnings where needed.


Former-commit-id: c5c17d5360c15bc88a169c4b6a1104393a88a219