Adding new CaloBaselineMonTool + Minor fix of CaloMonToolBase
CaloBaselineMonTool is a new tool to monitor the CaloCell baseline.
It is especially designed to:
- monitor the BCID subtraction within a bunch train
- monitor the pedestal in the empty bunches far from the trains
It is enabled by default in the Tier0 flow
Still to be done:
- trigger filtering
- specific switch off in useless stream

CaloMonToolBase is used a base class for many CaloMonitoring algs
The filtering based on LArCollisionTime was relevant only in CosmicCalo.
Adding a property to switch it on/off depending on the streams.
Jobopt still to be updated


Former-commit-id: 5e8dfb8d9e2b49aed52734e92b52e263f259df8b