Branch to update InDetGlobalMonitoring to use DataHandles, as discussed in ATLASRECTS-4027. This branch and MR complete the work done in !7779 and !3624.
In addition to moving to DataHandles, the branch also updates the declaration of members (using new compact form) and adds new configuation options, to allow jobs to run without having unmet dependencies on inputs that may not be present.

The reading in of some unused (and not normaly present) track collections is also removed, to simplify the code.


Former-commit-id: ccb2bfacbff7afa4e6e04f5621c9771f8f420d94