AthenaKernel: Preparations for allowing xAOD objects in metadata.
We want to allow DataLink<T> references to work in stores where the data
are actually stored as MetaCont<T>.
Introduce MetaContDataBucket, a specialization of DataBucket for MetaCont<T>.
This allows a conversion to the payload type T.
Also add a IProxyDict::sourceID, to return the correct metadata source
for the current event slot.
Also introduce METACONT_DEF macro, analogous to CONTCONT_DEF.
Add missing #includes to MetaCont.h.


Former-commit-id: f61da72e482a0cc2a88c09bdacda36b0e20d52be