CxxUtils: Fix race in ConcurrentRangeMap.
When a new version of map was installed, we were adding the old version
to the garbage list before updating the begin/last pointers.
This allows a race where data can be deleted under a client that's
still using it.  Swapping the order of these operations fixes this.

Also a few comment fixes and a tweak to the test code to reduce lock
contention.



Former-commit-id: ab5a5e3276f323e0e8c52d1e3e7d47badd2e2f6b