Add a WriteHandle for the CscDigitContainer to the CscDigitizationTool
The CscDigitContainer is now written by a WriteHandle, and the WriteHandles and WriteHandleKeys for both output containers are made global variables.
This is necessary because digitization can go through two completely independent paths.
Additionally, the configuration of the tool is changed to include the correct default value (previously the CscDigitContainer key was simply hard-coded in).
And the CSC overlay configuration is modified to allow the overlay to read the CscDigitContainer correctly.
Mentioning ATLASRECTS-4321 for bookkeeping purposes.


Former-commit-id: 9eaaa495323da15e31da4ac8a91766f87fabe1d0