Getting stopped particle simulation working
Stopped particle action is a stepping action, which needed a bit of
clean up in the user action config.  The cosmic filter config, which is
used for these jobs, also needed to be tweaked to not assume that the
job is for cosmics.  The preInclude for stopped particles needed an
update for the new name for the output file.


Former-commit-id: d329249d254ab27bac68882065aaaa5f4a61a364