Replace global buffer with a local one.
To avoid that concurrent users of one instance modify each other's
temporary storage which is used to store intermediate results,
the global buffer is replaced by a local one. To remove the
necessity for multiple allocations the result buffer is allocated to
also hold all the intermediate results and only shrunken to the final
size at the end of the computation.

This should address ATLASRECTS-4338.


Former-commit-id: 3e8649b908e7990f86d54014a097dc8389875d85