Make muon truth algs barcode-independent
This fix ensures that barcode shifts for the MuonTruthDecorationAlg, MuonTruthAssociationAlg, and MuonSegmentTruthAssociationAlg are correctly set in the configuration.
This resolves some of the muon issues reported in the validation of the MC18 truth strategy (see ATLPHYSVAL-553 for more details), owing to that strategy using different barcode shifts.
It also helps to address longstanding problems with muon truth matching not being barcode-independent (see ATLASRECTS-2962 for more details).
Tier0 test results are available at /afs/cern.ch/work/w/wleight/public/MuonSW/ATLASRECTS-2962/: no differences are observed, as expected.
The Tier0 test does claim a new memory leak in q431 AODtoTAG, but as there was a problem with the script during this part of it, and these changes don't affect AODtoTAG, I'm inclined not to believe in it.


Former-commit-id: 1d714251bd18cfbf0c39a9fbee9dfe40417e32be