Make muon truth algs barcode-independent
This fix ensures that barcode shifts for the MuonTruthDecorationAlg, MuonTruthAssociationAlg, and MuonSegmentTruthAssociationAlg are correctly set in the configuration.
This resolves some of the muon issues reported in the validation of the MC18 truth strategy (see ATLPHYSVAL-553 for more details), owing to that strategy using different barcode shifts.
It also helps to address longstanding problems with muon truth matching not being barcode-independent (see ATLASRECTS-2962 for more details).
This manual sweep resolves conflicts in the original sweep from 21.0 to master.


Former-commit-id: 8450ad170224300a72bb2e105f5791fc8bc3f321