Add support for online MVA TES
ATR-17501

This MR implements an online MVA calibration for tau triggers, similar to the offline MVA TES.
It is based solely on calorimeter information because it runs before tracking at the HLT.
Two new calo cluster moments, not used offline, but used online, have been added to the TauDefs.h enum.
The MvaTESEvaluator and MvaTESVariableDecorator have been split according to inTrigger(), as for other algorithms.
Some cleanup was made in the handling of mu (was stored as double, then converted to float), and nVtxPU
(was decorated in the variable decorator, then recomputed in the evaluator).
The algorithms have been imported in TrigTauAlgorithmsHolder to be able to define define triggers using MVA TES.
The actual triggers will be defined in a separate MR.


Former-commit-id: 256577b0f2b7dc46cc69d4b5fa7e3e931b60b709