Downgrading ERROR to WARNING for input files with different conditions in FileMetaDataTool
ATLASG-935

To avoid unnecessary crashes in MCORE DAOD production when some of the subprocesses produce empty files, this MR downgrades an ERROR to a WARNING for cases where the metadata conditions are not identical for input and output, as is the case when some of the input files are empty. Full details of this slightly convoluted problem can be found in the JIRA ticket.


Former-commit-id: a1e077edcfbe7b9b166d614faf2992ed75fc4d8c