Update tau dataset and code to build ART scripts from the makefile
Added a new data set for the tau processing, this is now used.

Added a a make target which builds only those art scripts that need to
be updated. This has to be done in a call to a script sadly, and can not
be done with make dependencies since the art scripts sadly need to change
the case so generating the target names wiht a pattern is non-trivial

The script reads the main xml driver file to decide which jobs are
relevant, and then loops over those jobs rebuilding only the ones which
are older than the relevant xml fragment


Former-commit-id: b425ccf898890b20c6eec26c9cdeb211f643ad41