Mechanism for blocking use of AllVariables in DAOD job options
https://indico.cern.ch/event/698346/contributions/2866069/attachments/1587652/2519126/DAOD_size_finalreport_weekly.pdf
https://indico.cern.ch/event/690171/ (minutes therein)

This MR puts into effect a decision of the OAB, on the suggestion of the DAOD task force, to block the use of AllVariables in the slimming helper for most physics DAODs. The mechnism is very simple - for any format asking for AllVariables, the slimming helper looks in a new list ```AllVariablesDisallowed.py``` and if the format is named in there, it returns an error message and crashes the job. The list of disallowed formats is initially empty but will be populated in the following months. (e.g. currently, the MR has no effect)


Former-commit-id: 660af6cd6d7f8abeea68209507c5d1b69fc4a3c1