Getting truth derivation making for HITS working again
Fixes one of the requests from ATLASSIM-3313.  Peter vG pointed out that
there were two instances of the same algorithm running with very
slightly different names.  Just protecting the second instance so that
it looks out for the first instance solves that.  Second issue was that
AODFix is scheduled to run even when we're running on EVNT and HITS
input, so I've disabled that in the derivations for EVNT input.  Now
this works!


Former-commit-id: ca85bdf4fbbac98af6cf57aad1c8482256975cf0