Add functionality to TruthIncident classes to pick up pre-defined decays
When we are running with quasi-stable particle simulation switched on it
is very useful to know what kind of interaction we are dealing with when
the `TruthSvc` comes to update the truth information. For this purpose the
interface of `ITruthIncident` has been extended to include an
`int interactionClassification() const` function. This returns zero in
the default implementation the function is then overloaded in
`Geant4TruthIncident` to use the the information from the current `G4Track`
to determine the nature of the interactions. There are four different
interaction classifications. Described as follows:

 - case 0: interaction of a particle without a pre-defined decay;

 - case 1: a particle with a pre-defined decay under-going a
           non-destructive interaction;

 - case 2: a particle with a pre-defined decay under-going a
           destructive interaction other than its pre-defined decay;

 - case 3: a particle under-going its pre-defined decay.

These will allow the TruthSvc to manipulate the `GenEvent` in different
ways in each case. This commit will help the work in ATLASSIM-3313.


Former-commit-id: cbb8cb5c71822c400bcc9565df9396a94fe7c4fc