Dropping unnecessary Geant4 UI command calls
During a discussion with the Geant4 experts about Geant4 UI commands
it was noted that the command `/geometry/navigator/check_mode true`
was being issued in `G4AtlasAlg` and `TransportTool`. Due to the point
in the initialization at which this is called it currently has no effect,
but if it were to be run it would greatly slow down simulation jobs
due to the extensive testing which it activates.

For this reason these calls are being dropped, to remove any possibility
of the code being activated in current/future production jobs.

See ATLSASIM-3454 for more discussion of how `G4UICommands` are used in
ATLAS simulation code.


Former-commit-id: 02b61a2426c21d04058d2035134fada217ba51a2