Add updateChildParticle method to ITruthIncident implementations
The new `updateChildParticle` method update the properties of a
child `HepMC::GenParticle` from a pre-defined interaction already
present in the `HepMC::GenEvent` based on the properties of the
ith child of the current `TruthIncident` (only used in quasi-stable
particle simulation). See ATLASSIM-3313.
Initially only the `Geant4TruthIncident` has a complete implementation
for this. The `ISFTruthIncident` just has a dummy implementation
as currently there are no plans to use quasi-stable particle simulation
with simulators other than Geant4.


Former-commit-id: 8c6f63b3826c246742f84fd1ad699e4e19433664