StoreGate: Work around scheduler issue with ReadDecorHandleKey and inheritance.
If, for example, @c D derives from @c B, and one uses
WriteDecorHandleKey<B>, then the scheduler will know only about the
B.DECOR output; it won't have a dependency for D.DECOR.  That means that
trying to use D.DECOR as a dependency downstream won't work.  To solve this,
we arrange that the CLID in the DataObjID for ReadDecorHandleKey<D>
is actually that for B; in that case, the scheduler will be happy.


Former-commit-id: 7f6f44fba60b6ef8a6c0a4cc243abb0165681d0f