Set the barcode offset for the MCTruthClassifier instance used by JetRec
For a number of years the barcodes of truth particles created during simulation
have started at 200001. Most downstream truth manipulation code assumes this
value and uses it to separate truth particles created in the simulation step
from those created in the generation step. There are some cases such as heavy
ion collisions where it is possible that the generator could create more than
200k particles though causing a problem. So in general it would be preferable
to set this offset to a higher value. Sometime ago changes were made to make
value configurable in the `MCTruthClassifier`, but this option was not
implemented in the configuration for most instances of the tool. This merge
request implements this configuration for the instance used by the Jet
reconstruction code. The derivations should be fixed properly in the `21.2`
branch though.

After the changes in this merge request and !8421 reconstruction shows
identical output in the cases that the offset is set to 200k or 1M (other
than the truth barcodes of course). Should fix ATLASRECTS-2962 and allow
for another iteration of ATLPHYSVAL-553.


Former-commit-id: 22ebc9eb04703adf827f7b9acf68a89b11d0b800