Enabling SUSY truth channel decoration for truth derivations
Adding the SUSY channel and PDG ID decorations to truth derivations.
Because of the python setup, this will only run on derivations that are
in fact SUSY signals.

At the same time, improving the logic by which mc channel number is
determined, to protect against bogus mc channel numbers and different
manners of writing job options.


Former-commit-id: 9688e56028c9c46af48312f8210e79d771ea96e5