remove the no longer supported ProofDriver
It was decided that ASG would no longer support PROOF (and the driver
was never functional in release 21.2), so I removed it now.  Besides
avoiding users getting confused this also means I no longer need to
maintain compatibility with the ProofDriver in other parts of
EventLoop, giving me some more freedom of how to organize things.


Former-commit-id: 0f7964a9994c0232ffe31d9b4261405fb0ad718d