Get rid of IBL flags in NoiseMapBuilder
With this commit backwards compatibility with Run-1 is broken, we now
assume IBL is there. This simplifies the structure of the code of the
NoiseMapBuilder, since we can just stay with a single mapping file
(PixelMapping_Run2.dat), and simplify the logic while iterating
over the pixel barrel layers: layer 0 is IBL, layer 1 is Blayer, etc.


Former-commit-id: 3428b10b6f28268bc1f38affea9ac71999b13989