Suppress warnings about invalid Identifiers for dead material
Some G4 volumes are partly active and particle dead depending on
where in the volume the energy was deposited. This means that it
is perfectly possible to get a dead material `LArG4Identifier` for
some parts of an otherwise active volume. In this case we should
not issue a warning message as the `Identifier` generated from the
`LArG4Identifier` is expected to be invalid.


Former-commit-id: 49bfecc9756ce6b3548ef1c8ab0922d0cf13519d