attempt to fix mis-match of translation/rotation deviation
The AlignmentDeviation class gets a new member which is a hash
to be created from the identifiers of all hits affected by the
deviation. Therefore it becomes much easier to check if the
two deviations affect the same hits but just comparing the hash.
Using this functionality a bug can be fixed more easily in the
RefitTool where the matching went wrong in case of deviations
affecting multiple chambers. One should investigate if this
hash cannot even be created earlier in the AlignmentTool and
used there. This would recude the number of for-loops.


Former-commit-id: 0f9d4c15e11d40589cab81ad7e2776d518490aa1