Fix bug in the type opf quotes used in the athena job options
All the ID Trigger bjet ART jobs are failing because of the wrong type of
quote used for the -c command optoin used for the command string.

Sadly the athena job shows no meaningful error so this took a while to
identify the issue. We have had no successful bjet tests for many weeks,
so this is rather an urgent request.


Former-commit-id: 7cd36984f3ed51527e24a153aaeccb3d8c830101