Transform support for handling the compression algorithm/level of temporary outputs (ATEAM-420)
This branch provides transform support for handling the compression algorithm and
compression level of the temporary outputs of AthenaMP. It sets the default
compression algorithm of RDO, ESD and AOD files to 2 (LZMA), and the compression
level to 1. For temporary outputs of multi-core sub-steps, it sets the compression
algorithm to 1 (ZLIB) and the compression level to 5.


Former-commit-id: ddff1d8eff0b63c49f50e3994e95c48ee1202ba5