Reduce the number of entries in the ttbar job which is timing out, add utility script
One job, the all_ttbar_pu80 job is taking too too long to processed for the art framework
so the number of events has been reduced.

Also a utility script have been added for the script generation stage (run only by the user
before making a merge request) to automatically post the jira request for any new data sets
to be added to the art container if they are not already there.

It would have been more efficienct, and quicker a nd generally better all round to just
add code to add the data sets directly rather than all this messing about with making a
request, waiting for someone to implement that request etc, but there you go.


Former-commit-id: 6dd09928138604687bc12196102d27b5b6b89af6