Declare undeclared tool handles.
Use in class ToolHandle initialisers to declare undeclared tool handles.
All these cases follow simple patterns, the tool handle is initialised
in the constructor, and in case of tool handle arrays the array is
initialised with exactly one element. The type of the tool without the
namespace, which is used in the initialisation, is used as property name.


Former-commit-id: accbfd50f7958bc247010fe25a4a159cad038c69