MuonSegmentMomentum: Fix assertion due to Eigen dimension mismatch.
We were getting an assertion failure on this line

    Amg::Vector3D resi(4),qoverp(4);

because we were trying to initialize three-vectors with a dimension of 4.

In fact, there is no real reason for these variables to be Eigen objects
in the first place ---- they're only used as arrays, not as vectors.
So just change them to C++ arrays.

Fixes crashs in the standard RecExCommon job seen in the dbg build.



Former-commit-id: 717d76e71d441a19c67101d65e082fd3726e2945