Making Calculator inherit as well from AsgMessaging.
Some internal reshuffling done, due to AsgMessaging being non-moveable and TrigGlobalEfficiencyCorrectionTool holding a vector of Calculators. Added one level of nesting so that TrigGlobalEfficiencyCorrectionTool now contains a single Calculator instance, and the latter a vector of the new structure Period (minimal wrapper of period boundaries + function pointer). One extra benefit is that the memory footprint is slightly reduced, as the duplication of caching resources between Calculators has been supressed.


Former-commit-id: b5acd10c1903125840cb3b12af6f0bd3ac8807a9