AthContainers: Use concurrent_vector for the vector of auxid information.
AuxTypeRegistry keeps a vector of information indexed by auxid.
Change this from a std::vector to a tbb::concurrent_vector.
This means that it can be read without needing to take out a lock.

Also remove the external locking functions from AuxTypeRegistry, and change
the lock from an upgradable mutex to an ordinary mutex.


Former-commit-id: 7bfa9de1337153d9b21793b375669f952b5c79ee