WIP: Changes to allow improve job processing in the IDTrigger art scripts
This request contains two main sets of related changes, plus some
minor fixes.

The minor fixes are just to update the data samples used by the RTT
and ART jobs and fix a bug on the command line in one of the RTT xml
files, which caused a rebuild of the global RTT configuration file
```
  TrigInDetValidation_TestConfiguration.xml
```
NB: this TrigInDetValidation_TestConfiguration.xml configuration file
    is very large and auto-generated - so please do not review this,
    as any general comments regarding this file will be ignored. In
    the future we may want to build this file automatically, although
    this is only for the RTT and so will be phased out reasonably soon.

The remaining two sets of changes are as follows:

1. minor changes to Tools/ART package to allow an art flag
   ```
   # art-ncores:
   ```
   to enable the art jobs to set how many cores to run on.

2. More extensive changes to the TrigInDetValidation package for
   the creation of the ART scripts.

None of these changes are used at P1 or at Tier 0 - they essentially
only affect the running of the jobs in the ART framework and are
not tested by any of the CI tests.

The changes to TrigInDetValidation appear extensive, but this is
misleading, since this package contains ~ 40 auto-generatted runscripts
for our ART jobs. By changing the structure of the autogeneration,
*all* of the run scripts are changed.

The runscripts (ie the files in TrigInDetValidation/test/test_\*.sh)
should *not* be reviewed as they are autogenerated and have all been
tested to be fully functional.

The significant changes are the modifications to the run script
generation to use a template file (utils/template.sh) for the
main part of the athena job running in the run scripts

This template file includes code to allow multiple athena jobs
to be run simultaneously, and as such this code is duplicated
in all the auto-generated test/test_\*.sh run scripts.

This template is used by the mkart-template script to build
the scripts.

There is also a new script to check that the datasets used by the
scripts are contained in the art container on the grid, and a script
to automatically post a jira request that any missing dataset be
added should it not already be present.

There are additional utilities to help calculate the number of files
etc that are needed for each job, plus some autogenerated configuration
files.

In addition the README file has been extensively updated.


Former-commit-id: 6c08c6dd87ae874cad8510f44f0185ad02d1669b