Adding contol for using empty BCIDs in a pattern for signal
This is an important hook for stopped particle or out of time decay
searches.  The option that I've added has a few settings: allowing
signal in any empty BCID, or allowing them in the first N BCIDs after a
filled crossing.  This should satisfy the use-cases I can think of, at
least.

In doing this, I had to actually copy the property value instead of
taking the const reference as had been done before.  That's in order to
be able to set it.  Another option would be to copy the values directly
into the member variable and do all the manipulations on that thing, but
that would be a bit more intrusive.


Former-commit-id: 6a88bbc718397957490191d6c8fdaa1dbbf80bf0