reducing CPU spent in RpcReadoutElement (ATR-16731)
There are a few events timing out for TrigMuSuperEF_MGonly.
Looking at them with Persint showed that there is high activity
in RPC chambers. Profiling the jobs indicated that the fit is
done very (probably too) often. During this procedure the
RpcReadoutElement is used a lot which for its part hammers on
the IdHelper to get again and again the same answer rather than
caching it. This merge request implements the caching (to reduce
the fits will be done later). The time saved is ~30% for this
particular events, but is likely not visible for others. The
frozen Tier0 tests passed without problems. Please, PROC, take
special care that the request does not violate any frozen Tier0
tests. It must not change the output by any means and if it does
there is a bug.


Former-commit-id: d12c8ffbee52c7b2c541ab07894ac1e3ba68bd32