Turn off Staco muon reconstruction for cosmics
This provides a fix for ATLASRECTS-4381 by disabling reconstruction of Staco muons in cosmic events.
The current recommendation is not to use Staco muons as studies indicate that they are mostly fakes, so this should not be an issue.
Since this change only affects cosmic data, the results of the Tier0 tests (available at /afs/cern.ch/work/w/wleight/public/MuonSW/ATLASRECTS-4381/RunTier0Tests.log) show no FT0 violation.
(This likely indicates an underlying problem with the Extrapolator, but as it is only an issue for cosmics, and there is nobody available to look at the Extraplator, this is the easier fix to implement.)


Former-commit-id: ee994a3adc364b2a729a091c5bfe84a2bd07766a