EventSelectorByteStream made to account for events already skipped in previous files (ATEAM-455)
The transform argument --skipEvents=N is meant to skip the first N events out of
the total number of events, summed over all the input files. However,
EventSelectorByteStream was not accounting for events already skipped in previous
files, resulting in N skipped events in every file in the input file list.


Former-commit-id: e3027dee7b8b625343a44edac573db3005fd9d72