Updating slepton configuration to allow more general application
The Sleptons G4Extension previously only allowed for three specific types of slepton according to a particular use case. The changes here allow for selectron/smuon/stau L+R combinations.

There is a new specialConfig in SimulationJobOptions to steer a new long-lived slepton simplified model, which makes use of the newly modified G4Extensions/Sleptons as well as the existing G4Extensions/Gauginos (with a minor modification to ensure the sparticle masses are picked up correctly).


Former-commit-id: 349ba86fdcf58a1778649de637ee947b9f6a8cb7