Updating slepton configuration to allow more general application
The Sleptons G4Extension previously only allowed for three specific types of slepton according to a particular use case. The changes here allow for selectron/smuon/stau L+R combinations.

There is a new specialConfig in SimulationJobOptions to steer a new long-lived slepton simplified model, which makes use of the newly modified G4Extensions/Sleptons as well as the existing G4Extensions/Gauginos (with a minor modification to ensure the sparticle masses are picked up correctly).


Former-commit-id: 2852c731bd71a65ef7b8309442c4e8fddf26cfaf