Support G4VPhysicsConstructor usage for Geant4 >= 10.3
Prior to Geant4 10.3, G4VPhysicsConstructor provided direct access to
the particle iterator via a #define. In 10.3, this was removed and
replaced with a requirement to call a member function.

Update clients of the #define to use new member function call. Retain
existing preprocessor symbols, add new definitions for latest G4
versions that are just the member function call on 'this'.


Former-commit-id: b8ea00db03cbd57876b63dfb78a89d7ee5bac3dc