Fix for ATLASRECTS-4099
This fixes a bug in which hits in clusters were not being properly accounted for in the ambiguity resolution.
After some further discussion, it was decided to modify the ambiguity resolution to only take into account precision hits.
Changes to the PRD_AssociationTool allow it to correctly identify hits in clusters and to only accept precision muon hits.
(The changes will only apply to muon hits so that tracking is not affected.)
Changes to the MuonAmbiTrackSelectionTool cause it to only accept precision hits.


Former-commit-id: 6241c9d4d1d4697d68351e81c22ef18f82ff764e