Fix some missing track collections in the FTK analyses, fix the chains for the MinBias analyses
Some of the track collections to be used for the FTK analyses were omitted in error and these have now been added .

In addition, the analyses for teh minbias analyses have been updated so that the MinBias analyses should now work correctly.

In addition a small change to the formating of the time stamps has been made to the art template  file, and this mean that the art scripts have all been rebuilt.


Former-commit-id: d70b41f33d3a13aeefa4df78c8403c4c695d5b8b