Robustness improvements for the TruthSvc
It was found that calls to `ITruthIncident::parentParticleAfterIncident(...)`
may alter the output of subsequent calls to `ITruthIncident::parentParticle()`.
(Whether this should be the case is a separate issue to be followed up
elsewhere.)

The `TruthSvc` needed a couple of tweaks to make it robust against
this feature.  At the same time some of the logic was tidied up to
make it cleaner and more descriptive comments added in a couple of
places. See ATLASSIM-3313.


Former-commit-id: 147a1d97c669f1f5925dafd0b5d29433e3c88fa4