Prevent unit tests running the same transform from running in parallel (ATLINFR-2296)
In ATLINFR-2296 it was noted that the CI attempts to run unit tests in
parallel.  As unit tests all run in the same directory this causes
problems if two tests try to write to the same file(s), as is the
case with the simualtion configuration unit tests, where `AtlasG4_tf`
and `Sim_tf` are run in mulitple configurations in different tests.

By using the `RESOURCE_LOCK` property it is possible to prevent
unit tests running the same job transform from running in parallel.
This allows all unit tests to succeed even when running in parallel
in local tests.


Former-commit-id: f4928e513e37e40231830179c6e4d20d58e6f273