Change default python logging level to INFO
Change the default log level of the Logging module from WARNING to INFO.
This has zero impact when used within athena.py as the default log level
of athena.py is INFO anyway. It will have an impact when used from plain
python. But since our developers are used to seeing INFO messages by
default this should be a better default value (although it's no longer
consistent with the default of the python logging library).


Former-commit-id: 0b65559a8fcd02b761b8be286774fb86b97b1323