Bugfix for 2017 tau triggers with BDT tau ID
It was found (ATR-17789,ATR-17771) that during the implementation of 2018 tau triggers, the counts of 2017 tau triggers had significantly dropped.
This was traced to the retrieval of mu in TauIDVarCalculator, which was moved from execute() to eventInitialize(), to be performed once per event.
However, the piece of tau trigger code which takes care of the BDT evaluation did not call eventInitialize, which resulted in a null default value for mu.
And the BDT cut is a function of tau pt and mu...
TrigTauDiscriBuilder now calls eventInitialize and eventFinalize.
In TauIDVarCalculator, the retrival of mu is done via EventInfo and not via TauEventData, in order to save the implementation of a TauEventData layer in TrigTauDiscriBuilder.
Because this TauEventData would have been used only in TauIDVarCalculator for the mu retrieval, which is overkill and we then save a bit of memory...
This MR also changes algorithm names (prefix), which were set to be identical for 2017 and 2018 chains, by mistake.
Finally, 2 ROOT files used for RNN ID flattening were updated in TrigTauAlgorithmsHolder.py.
Their transfer to cvmfs has been requested (ATLINFR-2299) but is not yet effective.
The previous ROOT files were not closed properly. The content is identical.


Former-commit-id: b174be85ccbea158980f24e02ff7efb67e87c2e2